+++ libretroshare/src/chat/p3chatservice.h
@@ -110,7 +110,7 @@ public:
                                                    int contactType, int groupChatType, std::string rsPeerIdStr, ChatLobbyId chatLobbyId, std::string uId);
     virtual std::vector<conversationInfo> getConversationItemList();
 
-    virtual void updateRecentTimeOfItemInConversationList(std::string uId, std::string nickInGroupChat, unsigned int lastMsgDatetime, std::string textmsg, bool isOtherMsg );
+    virtual void updateRecentTimeOfItemInConversationList(std::string uId, std::string nickInGroupChat, long long lastMsgDatetime, std::string textmsg, bool isOtherMsg );
     virtual void sortConversationItemListByRecentTime();
     virtual void updateUnreadNumberOfItemInConversationList(std::string uId, unsigned int unreadNumber, bool isReset);
     virtual std::string getSeletedUIdBeforeSorting(int row);
