+++ libretroshare/src/rsserver/p3msgs.h
@@ -141,7 +141,7 @@ class p3Msgs: public RsMsgs
                                                          unsigned int UnreadMessagesCount, unsigned int lastMsgDatetime, std::string lastMessage, bool isOtherLastMsg,
                                                          int contactType, int groupChatType, std::string rsPeerIdStr, ChatLobbyId chatLobbyId, std::string uId);
           virtual std::vector<conversationInfo> getConversationItemList();
-          virtual void updateRecentTimeOfItemInConversationList(std::string uId, std::string nickInGroupChat, unsigned int lastMsgDatetime, std::string textmsg, bool isOtherMsg );
+          virtual void updateRecentTimeOfItemInConversationList(std::string uId, std::string nickInGroupChat, long long lastMsgDatetime, std::string textmsg, bool isOtherMsg );
           virtual void sortConversationItemListByRecentTime();
           virtual void updateUnreadNumberOfItemInConversationList(std::string uId, unsigned int unreadNumber, bool isReset);
           virtual std::string getSeletedUIdBeforeSorting(int row);
